# mesh scalar renderer optimization