this PR speeds up mesh scalar rendering.
Mesh scalar rendering is slower when there is a zoom on faces, especially under Windows.

Tested for one rendering full screen : 

* under windows: 
 before: # 350 ms, 
 after: # 140 ms

* under ubuntu 20.04:
 before: # 150 ms
 after: # 120 ms