@nyalldawson 2021-01-04
Backport to 3.16 as 49bba32899f39e382316a686e27e29f274033f9f